refactor: faz PanelSelector conformar com o padrão de código
